Initial commit
This repository begins with all the modules used by the image
sonification script because they are not integrated yet in a library.
The principal script 'img_sonif.py' is the principal purpose of this
repository, it contains the process of image sonification developed to
be integrated into the sonoUno interface.